+++ testutil/network/network.go
@@ -22,15 +22,18 @@ import (
 	"github.com/tendermint/tendermint/node"
 	tmclient "github.com/tendermint/tendermint/rpc/client"
 	dbm "github.com/tendermint/tm-db"
+	"google.golang.org/grpc"
 
 	"github.com/cosmos/cosmos-sdk/baseapp"
 	"github.com/cosmos/cosmos-sdk/client"
 	clientkeys "github.com/cosmos/cosmos-sdk/client/keys"
+	"github.com/cosmos/cosmos-sdk/client/tx"
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
 	"github.com/cosmos/cosmos-sdk/server"
 	"github.com/cosmos/cosmos-sdk/server/api"
 	srvconfig "github.com/cosmos/cosmos-sdk/server/config"
+	servertypes "github.com/cosmos/cosmos-sdk/server/types"
 	"github.com/cosmos/cosmos-sdk/simapp"
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -45,9 +48,9 @@ var lock = new(sync.Mutex)
 
 // AppConstructor defines a function which accepts a network configuration and
 // creates an ABCI Application to provide to Tendermint.
-type AppConstructor = func(val Validator) server.Application
+type AppConstructor = func(val Validator) servertypes.Application
 
-func NewSimApp(val Validator) server.Application {
+func NewSimApp(val Validator) servertypes.Application {
 	return simapp.NewSimApp(
 		val.Ctx.Logger, dbm.NewMemDB(), nil, true, make(map[int64]bool), val.Ctx.Config.RootDir, 0,
 		baseapp.SetPruning(storetypes.NewPruningOptionsFromString(val.AppConfig.Pruning)),
@@ -59,6 +62,7 @@ func NewSimApp(val Validator) server.Application {
 // in-process local testing network.
 type Config struct {
 	Codec            codec.Marshaler
+	LegacyAmino      *codec.Codec
 	TxConfig         client.TxConfig
 	AccountRetriever client.AccountRetriever
 	AppConstructor   AppConstructor             // the ABCI application constructor
@@ -85,6 +89,7 @@ func DefaultConfig() Config {
 	return Config{
 		Codec:            encCfg.Marshaler,
 		TxConfig:         encCfg.TxConfig,
+		LegacyAmino:      encCfg.Amino,
 		AccountRetriever: authtypes.NewAccountRetriever(encCfg.Marshaler),
 		AppConstructor:   NewSimApp,
 		GenesisState:     simapp.ModuleBasics.DefaultGenesis(encCfg.Marshaler),
@@ -118,7 +123,7 @@ type (
 		BaseDir    string
 		Validators []*Validator
 
-		config Config
+		Config Config
 	}
 
 	// Validator defines an in-process Tendermint validator node. Through this object,
@@ -141,6 +146,7 @@ type (
 
 		tmNode *node.Node
 		api    *api.Server
+		grpc   *grpc.Server
 	}
 )
 
@@ -157,7 +163,7 @@ func New(t *testing.T, cfg Config) *Network {
 		T:          t,
 		BaseDir:    baseDir,
 		Validators: make([]*Validator, cfg.NumValidators),
-		config:     cfg,
+		Config:     cfg,
 	}
 
 	t.Log("preparing test network...")
@@ -203,6 +209,11 @@ func New(t *testing.T, cfg Config) *Network {
 			rpcAddr, _, err := server.FreeTCPAddr()
 			require.NoError(t, err)
 			tmCfg.RPC.ListenAddress = rpcAddr
+
+			_, grpcPort, err := server.FreeTCPAddr()
+			require.NoError(t, err)
+			appCfg.GRPC.Address = fmt.Sprintf("0.0.0.0:%s", grpcPort)
+			appCfg.GRPC.Enable = true
 		}
 
 		logger := log.NewNopLogger()
@@ -278,16 +289,21 @@ func New(t *testing.T, cfg Config) *Network {
 		require.NoError(t, err)
 
 		memo := fmt.Sprintf("%s@%s:%s", nodeIDs[i], p2pURL.Hostname(), p2pURL.Port())
-		tx := authtypes.NewStdTx([]sdk.Msg{createValMsg}, authtypes.StdFee{}, []authtypes.StdSignature{}, memo) //nolint:staticcheck // SA1019: authtypes.StdFee is deprecated
-		txBldr := authtypes.TxBuilder{}.
+		txBuilder := cfg.TxConfig.NewTxBuilder()
+		require.NoError(t, txBuilder.SetMsgs(createValMsg))
+		txBuilder.SetMemo(memo)
+
+		txFactory := tx.Factory{}
+		txFactory = txFactory.
 			WithChainID(cfg.ChainID).
 			WithMemo(memo).
-			WithKeybase(kb)
+			WithKeybase(kb).
+			WithTxConfig(cfg.TxConfig)
 
-		signedTx, err := txBldr.SignStdTx(nodeDirName, tx, false)
+		err = tx.Sign(txFactory, nodeDirName, txBuilder)
 		require.NoError(t, err)
 
-		txBz, err := cfg.Codec.MarshalJSON(signedTx)
+		txBz, err := cfg.TxConfig.TxJSONEncoder()(txBuilder.GetTx())
 		require.NoError(t, err)
 		require.NoError(t, writeFile(fmt.Sprintf("%v.json", nodeDirName), gentxsDir, txBz))
 
@@ -298,6 +314,7 @@ func New(t *testing.T, cfg Config) *Network {
 			WithHomeDir(tmCfg.RootDir).
 			WithChainID(cfg.ChainID).
 			WithJSONMarshaler(cfg.Codec).
+			WithCodec(cfg.LegacyAmino).
 			WithTxConfig(cfg.TxConfig).
 			WithAccountRetriever(cfg.AccountRetriever)
 
@@ -424,7 +441,7 @@ func (n *Network) Cleanup() {
 		}
 	}
 
-	if n.config.CleanupDir {
+	if n.Config.CleanupDir {
 		_ = os.RemoveAll(n.BaseDir)
 	}